LayoutTests/tiled-drawing/scrolling/scroll-snap/scroll-snap-mandatory-then-proximity-expected.txt

Scrolling in container with scroll-snap-type: y mandatory
- Did the scrolling snap to the top? NO
- Did scrolling snap to the second box? YES
Scrolling in container with scroll-snap-type: y proximity
- Did the scrolling snap to the top? NO
- Did scrolling snap to the second box? NO
